app: Support /etc/hosts in writable path
With [1] coming up, it is not necessary to mount the rootfs read-writable anymore,
and blocklists will be stored persistently. In order to accommodate for both old
and new approaches we move the temporary and backup files to /var/lib/misc/uAdBlockNG
and check for successful enablement using "cmp". Whether the running system supports
/etc/hosts in writable path is checked via the "mountpoint" command.

1: https://gitlab.com/ubports/development/core/hybris-support/lxc-android-config/-/merge_requests/171